backend/agents/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2
google-generativeai==0.8.2
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==10.4.0
pytesseract==0.3.13
speechrecognition==3.10.4
pydub==0.25.1
openai-whisper==20231117
openai>=1.0.0
# InLegalBERT + FAISS dependencies
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0


